requirements.txt
aws-cdk-lib==2.62.2
constructs>=10.0.0,<11.0.0
aws-solutions-constructs.aws-kinesis-firehose_s3>=2.25.0
boto3==1.16.26